group joints by profile instead of joints
It's *almost never* useful to visualize position, velocity, and acceleration
for more than one joint at a time.

On the other hand, it is *very* useful to plot all positions,
*or* all velocities, *or* all acceleration for all joints in a trajectory
together to inspect the dynamics profile and e.g. look for the bottleneck
joint of a trajectory.

While the former layout made the first described use easy, the inverted
hierarchy here makes the second use-case only one click.